Lab1_AI_Search_RAG_pipeline/requirement.txt.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==2.15.8
blinker==1.6.2
boltons==23.0.0
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
conda==23.5.2
conda-content-trust==0.1.3
conda-libmamba-solver==23.5.0
conda-package-handling==2.1.0
conda_package_streaming==0.8.0
contourpy==1.1.1
cryptography==39.0.1
cycler==0.11.0
dill==0.3.7
distro==1.9.0
Flask==2.3.3
fonttools==4.42.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.10.0
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
libmambapy==1.4.1
MarkupSafe==2.1.3
matplotlib==3.8.0
mccabe==0.7.0
menuinst==1.4.19
numpy==1.26.0
openai==1.65.2
packaging==23.0
Pillow==10.0.1
pip==23.1.2
platformdirs==3.10.0
pluggy==1.0.0
pycosat==0.6.4
pycparser==2.21
pydantic==2.11.5
pydantic_core==2.33.2
pylint==2.17.6
pyOpenSSL==23.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.29.0
ruamel.yaml==0.17.21
setuptools==67.8.0
six==1.16.0
sniffio==1.3.1
tomlkit==0.12.1
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.14.0
typing-inspection==0.4.1
urllib3==1.26.16
Werkzeug==2.3.7
wheel==0.38.4
win-inet-pton==1.1.0
wrapt==1.15.0
zstandard==0.19.0